reconstruct codes
