+++ examples/pytorch/wan/transformer_wan_flashinfer.py
@@ -28,1160 +28,35 @@ standard RoPE APIs.
 
 import logging
 import math
-import warnings
+import os
+import sys
+from contextlib import nullcontext
 from dataclasses import dataclass
-from enum import Enum
-from typing import Any, Dict, Literal, Optional, Tuple, Union
+from pathlib import Path
+from typing import Any, Literal, Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-import flashinfer
-from flashinfer.prefill import (
-    BatchPrefillWithRaggedKVCacheWrapper,
-    trtllm_batch_context_with_kv_cache,
-)
-from flashinfer.cudnn import cudnn_batch_prefill_with_kv_cache
-from flashinfer.gemm import (
-    gemm_fp8_nt_groupwise,
-    gemm_fp8_nt_blockscaled,
-    batch_deepgemm_fp8_nt_groupwise,
-)
-from flashinfer.utils import get_compute_capability
-
 logger = logging.getLogger(__name__)
 
-# Default workspace buffer size for FlashInfer batched attention kernels.
-# 128 MB is the recommended size in FlashInfer documentation; sufficient for
-# typical Wan video transformer workloads (up to ~26K visual tokens).
-_DEFAULT_WORKSPACE_SIZE = 128 * 1024 * 1024
 
-
-class GEMMBackend(Enum):
-    """GEMM backend selection for FlashInferLinear."""
-
-    TORCH = "torch"  # Standard PyTorch (fallback, any GPU)
-    BF16 = "bf16"  # FlashInfer mm_bf16 (SM100+)
-    FP8 = "fp8"  # FlashInfer mm_fp8 with TRT-LLM backend (SM89+)
-    FP8_SM90 = "fp8_sm90"  # FlashInfer fp8_blockscale_gemm_sm90 (SM90 only)
-    BMM_FP8 = "bmm_fp8"  # FlashInfer bmm_fp8 with cublas backend (SM89+)
-    FP8_GROUPWISE = "fp8_groupwise"  # FlashInfer gemm_fp8_nt_groupwise (SM100+)
-    FP8_BLOCKSCALED = "fp8_blockscaled"  # FlashInfer gemm_fp8_nt_blockscaled (SM100+)
-    BATCH_DEEPGEMM_FP8 = (
-        "batch_deepgemm_fp8"  # FlashInfer batch_deepgemm_fp8_nt_groupwise (SM100/SM103)
-    )
-    FP4 = "fp4"  # FlashInfer mm_fp4 (SM100+)
-    BMM_BF16 = "bmm_bf16"  # FlashInfer bmm_bf16 (SM100+)
-    MXFP8 = "mxfp8"  # FlashInfer mm_mxfp8 (SM100+/SM120+)
-    BMM_MXFP8 = "bmm_mxfp8"  # FlashInfer bmm_mxfp8 (SM100+/SM120+)
-
-
-def _check_gemm_backend_support(backend: GEMMBackend, device: torch.device) -> bool:
-    """Check if a GEMM backend is supported on the current device."""
-    if backend == GEMMBackend.TORCH:
-        return True
-
-    major, minor = get_compute_capability(device)
-    sm = major * 10 + minor
-
-    if backend == GEMMBackend.BF16:
-        # mm_bf16 requires SM100+ (Blackwell)
-        return sm >= 100
-    elif backend == GEMMBackend.FP8:
-        # mm_fp8 (TRT-LLM low-latency) requires SM89+
-        # Note: On SM90, prefer FP8_SM90 (fp8_blockscale_gemm_sm90) as it's more reliable
-        # TRT-LLM has strict dimension constraints that may cause failures
-        return sm >= 89 and sm != 90  # Disable on SM90, use FP8_SM90 instead
-    elif backend == GEMMBackend.FP8_SM90:
-        # fp8_blockscale_gemm_sm90 requires exactly SM90
-        return sm == 90
-    elif backend == GEMMBackend.BMM_FP8:
-        # bmm_fp8 with cublas backend works on SM89+
-        return sm >= 89
-    elif backend == GEMMBackend.FP8_GROUPWISE:
-        # gemm_fp8_nt_groupwise requires SM100+ (Blackwell)
-        return sm >= 100
-    elif backend == GEMMBackend.FP8_BLOCKSCALED:
-        # gemm_fp8_nt_blockscaled requires SM100+ (Blackwell)
-        return sm >= 100
-    elif backend == GEMMBackend.BATCH_DEEPGEMM_FP8:
-        # batch_deepgemm_fp8_nt_groupwise requires SM100 or SM103
-        return sm in (100, 103)
-    elif backend == GEMMBackend.FP4:
-        # mm_fp4 requires SM100+ (Blackwell)
-        return sm >= 100
-    elif backend == GEMMBackend.BMM_BF16:
-        # bmm_bf16 requires SM100+ (Blackwell)
-        return sm >= 100
-    elif backend == GEMMBackend.MXFP8:
-        # mm_mxfp8 requires SM100+ or SM120+
-        return sm >= 100
-    elif backend == GEMMBackend.BMM_MXFP8:
-        # bmm_mxfp8 requires SM100+ or SM120+
-        return sm >= 100
-
-    return False
-
-
-class FlashInferLinear(nn.Module):
-    """
-    Linear layer using FlashInfer's optimized GEMM kernels.
-
-    This module provides drop-in replacement for nn.Linear with support for
-    multiple precision backends:
-    - TORCH: Standard PyTorch (works on any GPU)
-    - BF16: FlashInfer mm_bf16 with CUTLASS/cuDNN (SM100+)
-    - FP8: FlashInfer mm_fp8 with TRT-LLM low-latency backend (SM89+)
-    - FP4: FlashInfer mm_fp4 with block quantization (SM100+)
-
-    Args:
-        in_features: Size of input features
-        out_features: Size of output features
-        bias: Whether to include bias (only supported with TORCH backend)
-        backend: GEMM backend to use (see GEMMBackend enum for options)
-        device: Device to place the module on
-        dtype: Data type for weights (for TORCH/BF16 backends)
-        online_act_quant: True for online activation scale computation, False for fixed default scale
-    """
-
-    def __init__(
-        self,
-        in_features: int,
-        out_features: int,
-        bias: bool = True,
-        backend: Literal[
-            "torch",
-            "bf16",
-            "fp8",
-            "fp8_sm90",
-            "bmm_fp8",
-            "fp8_groupwise",
-            "fp8_blockscaled",
-            "batch_deepgemm_fp8",
-            "fp4",
-            "bmm_bf16",
-            "mxfp8",
-            "bmm_mxfp8",
-        ] = "torch",
-        device: Optional[torch.device] = None,
-        dtype: Optional[torch.dtype] = None,
-        online_act_quant: bool = True,
-    ):
-        super().__init__()
-        self.in_features = in_features
-        self.out_features = out_features
-        self.has_bias = bias
-        self.online_act_quant = online_act_quant
-
-        # Resolve device — FlashInfer kernels require CUDA. We accept None (default
-        # to current CUDA device) but reject CPU since none of the GEMM backends
-        # except TORCH would actually function there.
-        if device is None:
-            if not torch.cuda.is_available():
-                raise RuntimeError(
-                    "FlashInferLinear requires CUDA, but no CUDA device is available."
-                )
-            device = torch.device("cuda")
-        elif torch.device(device).type != "cuda":
-            raise RuntimeError(
-                f"FlashInferLinear requires a CUDA device, got {device!r}."
-            )
-        self.device = device
-
-        self._backend = GEMMBackend(backend)
-
-        # Validate backend support
-        if not _check_gemm_backend_support(self._backend, device):
-            warnings.warn(
-                f"{self._backend.value} GEMM backend is not supported on this device; "
-                "falling back to TORCH.",
-                RuntimeWarning,
-                stacklevel=2,
-            )
-            self._backend = GEMMBackend.TORCH
-
-        # Check dimension constraints for FP8 backends
-        if self._backend == GEMMBackend.FP8_SM90:
-            # fp8_blockscale_gemm_sm90 requires: N % 64 == 0, K % 128 == 0
-            if out_features % 64 != 0 or in_features % 128 != 0:
-                warnings.warn(
-                    f"FP8_SM90 requires N%64==0 and K%128==0, got N={out_features}, "
-                    f"K={in_features}; falling back to TORCH.",
-                    RuntimeWarning,
-                    stacklevel=2,
-                )
-                self._backend = GEMMBackend.TORCH
-
-        # Check dimension constraints for MXFP8 backends
-        if self._backend in (GEMMBackend.MXFP8, GEMMBackend.BMM_MXFP8):
-            # MXFP8 uses block size 32, so K must be divisible by 32
-            if in_features % 32 != 0:
-                warnings.warn(
-                    f"MXFP8 requires K%32==0, got K={in_features}; falling back to TORCH.",
-                    RuntimeWarning,
-                    stacklevel=2,
-                )
-                self._backend = GEMMBackend.TORCH
-
-        # Check dimension constraints for NT groupwise/blockscaled FP8
-        if self._backend in (
-            GEMMBackend.FP8_GROUPWISE,
-            GEMMBackend.FP8_BLOCKSCALED,
-            GEMMBackend.BATCH_DEEPGEMM_FP8,
-        ):
-            if in_features % 128 != 0:
-                warnings.warn(
-                    f"{self._backend.value} requires K%128==0, got K={in_features}; "
-                    "falling back to TORCH.",
-                    RuntimeWarning,
-                    stacklevel=2,
-                )
-                self._backend = GEMMBackend.TORCH
-
-        # Handle bias - only TORCH backend supports bias directly
-        # Note: bias is added separately in forward() for non-TORCH backends
-
-        # Initialize weights based on backend
-        if self._backend == GEMMBackend.FP8:
-            self._init_fp8_weights(dtype or torch.bfloat16)
-        elif self._backend == GEMMBackend.FP8_SM90:
-            self._init_fp8_sm90_weights(dtype or torch.bfloat16)
-        elif self._backend == GEMMBackend.BMM_FP8:
-            self._init_bmm_fp8_weights(dtype or torch.bfloat16)
-        elif self._backend in (
-            GEMMBackend.FP8_GROUPWISE,
-            GEMMBackend.FP8_BLOCKSCALED,
-            GEMMBackend.BATCH_DEEPGEMM_FP8,
-        ):
-            self._init_fp8_nt_weights(dtype or torch.bfloat16)
-        elif self._backend == GEMMBackend.FP4:
-            self._init_fp4_weights(dtype or torch.bfloat16)
-        elif self._backend == GEMMBackend.MXFP8:
-            self._init_mxfp8_weights(dtype or torch.bfloat16)
-        elif self._backend == GEMMBackend.BMM_MXFP8:
-            self._init_bmm_mxfp8_weights(dtype or torch.bfloat16)
-        else:
-            # TORCH, BF16, and BMM_BF16 use standard weight layout
-            self._init_standard_weights(dtype or torch.bfloat16)
-
-    def _init_standard_weights(self, dtype: torch.dtype):
-        """Initialize weights for TORCH/BF16 backends."""
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        # Kaiming initialization
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # For BF16 backend, cache the transposed weight
-        self._weight_t_cache = None
-        self._fp8_prepared = False
-
-    def _init_fp8_weights(self, dtype: torch.dtype):
-        """Initialize weights for FP8 backend."""
-        # Store original weight for conversion
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # FP8 specific attributes (will be computed during first forward or prepare_weights)
-        self.register_buffer("_weight_fp8", None)
-        self.register_buffer("_weight_scale", None)
-        self._permutation_cache: Dict[torch.Size, torch.Tensor] = {}
-        self._fp8_prepared = False
-
-    def _init_fp4_weights(self, dtype: torch.dtype):
-        """Initialize weights for FP4 backend."""
-        # Store original weight for conversion
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # FP4 specific attributes
-        self.register_buffer("_weight_fp4", None)
-        self.register_buffer("_weight_descale", None)
-        self._fp4_prepared = False
-
-    def _init_fp8_sm90_weights(self, dtype: torch.dtype):
-        """Initialize weights for FP8_SM90 backend (fp8_blockscale_gemm_sm90)."""
-        # Store original weight - will be quantized to FP8 with block scales
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # FP8 SM90 specific attributes (128x128 block scales)
-        self.register_buffer("_weight_fp8_sm90", None)
-        self.register_buffer("_weight_scale_sm90", None)
-        self._fp8_sm90_prepared = False
-
-    def _init_bmm_fp8_weights(self, dtype: torch.dtype):
-        """Initialize weights for BMM_FP8 backend (bmm_fp8 with cublas)."""
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # BMM FP8 specific attributes
-        self.register_buffer("_weight_fp8_bmm", None)
-        self.register_buffer("_weight_scale_bmm", None)
-        self._bmm_fp8_prepared = False
-
-    def _init_fp8_nt_weights(self, dtype: torch.dtype):
-        """Initialize weights for FP8_GROUPWISE / FP8_BLOCKSCALED / BATCH_DEEPGEMM_FP8 backends.
-
-        These use NT-format GEMM: a(m,k) @ b(n,k)^T with block-scale quantization.
-        """
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # NT FP8 specific attributes (weight stored as (N, K) in FP8)
-        self.register_buffer("_weight_fp8_nt", None)
-        self.register_buffer("_weight_scale_nt", None)
-        self._fp8_nt_prepared = False
-
-    def _init_mxfp8_weights(self, dtype: torch.dtype):
-        """Initialize weights for MXFP8 backend (mm_mxfp8)."""
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # MXFP8 specific attributes (block size 32)
-        self.register_buffer("_weight_mxfp8", None)
-        self.register_buffer("_weight_scale_mxfp8", None)
-        self._mxfp8_prepared = False
-
-    def _init_bmm_mxfp8_weights(self, dtype: torch.dtype):
-        """Initialize weights for BMM_MXFP8 backend (bmm_mxfp8)."""
-        self.weight = nn.Parameter(
-            torch.empty(
-                self.out_features, self.in_features, device=self.device, dtype=dtype
-            )
-        )
-        if self.has_bias:
-            self.bias = nn.Parameter(
-                torch.zeros(self.out_features, device=self.device, dtype=dtype)
-            )
-        else:
-            self.register_parameter("bias", None)
-
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-        if self.bias is not None:
-            fan_in = self.in_features
-            bound = 1 / math.sqrt(fan_in) if fan_in > 0 else 0
-            nn.init.uniform_(self.bias, -bound, bound)
-
-        # BMM MXFP8 specific attributes
-        self.register_buffer("_weight_mxfp8_bmm", None)
-        self.register_buffer("_weight_scale_mxfp8_bmm", None)
-        self._bmm_mxfp8_prepared = False
-
-    @torch.no_grad()
-    def prepare_weights(self):
-        """Prepare weights for optimized inference (call after loading weights)."""
-        if self._backend == GEMMBackend.FP8:
-            self._prepare_fp8_weights()
-        elif self._backend == GEMMBackend.FP8_SM90:
-            self._prepare_fp8_sm90_weights()
-        elif self._backend == GEMMBackend.BMM_FP8:
-            self._prepare_bmm_fp8_weights()
-        elif self._backend in (
-            GEMMBackend.FP8_GROUPWISE,
-            GEMMBackend.FP8_BLOCKSCALED,
-            GEMMBackend.BATCH_DEEPGEMM_FP8,
-        ):
-            self._prepare_fp8_nt_weights()
-        elif self._backend == GEMMBackend.FP4:
-            self._prepare_fp4_weights()
-        elif self._backend == GEMMBackend.BF16:
-            self._prepare_bf16_weights()
-        elif self._backend == GEMMBackend.BMM_BF16:
-            self._prepare_bmm_bf16_weights()
-        elif self._backend == GEMMBackend.MXFP8:
-            self._prepare_mxfp8_weights()
-        elif self._backend == GEMMBackend.BMM_MXFP8:
-            self._prepare_bmm_mxfp8_weights()
-
-    @torch.no_grad()
-    def _prepare_bf16_weights(self):
-        """Prepare weights for BF16 backend (pre-transpose)."""
-        # mm_bf16 expects weight in column-major layout (k, n)
-        # Our weight is (out_features, in_features) = (n, k), need to transpose to (k, n)
-        self._weight_t_cache = self.weight.data.t().contiguous()
-
-    @torch.no_grad()
-    def _prepare_fp8_weights(self):
-        """Prepare weights for FP8 backend."""
-        if self._fp8_prepared:
-            return
-
-        # Convert weight to FP8
-        weight = self.weight.data
-        finfo = torch.finfo(torch.float8_e4m3fn)
-
-        # Compute scale
-        amax = weight.abs().max().clamp(min=1e-12)
-        scale = finfo.max / amax
-
-        # Quantize to FP8
-        weight_scaled = (weight * scale).clamp(finfo.min, finfo.max)
-        weight_fp8 = weight_scaled.to(torch.float8_e4m3fn)
-
-        # Prepare for TRT-LLM low latency GEMM
-        # Weight shape: (n, k) -> needs to be (k // block_size, n, block_size)
-        prepared_weight = flashinfer.prepare_low_latency_gemm_weights(
-            weight_fp8, self._permutation_cache
-        )
-
-        self._weight_fp8 = prepared_weight
-        self._weight_scale = 1.0 / scale
-        self._fp8_prepared = True
-
-    @torch.no_grad()
-    def _prepare_fp4_weights(self):
-        """Prepare weights for FP4 backend."""
-        if self._fp4_prepared:
-            return
-
-        # Use FlashInfer's nvfp4_quantize for FP4 conversion
-        weight = self.weight.data
-        # nvfp4_quantize expects (m, k) layout for row-major
-        # Our weight is (out_features, in_features) = (n, k)
-
-        # For mm_fp4, weight needs to be column-major (k, n)
-        weight_t = weight.t().contiguous()  # (k, n)
-
-        # Quantize to NV-FP4
-        weight_fp4, weight_descale = flashinfer.nvfp4_quantize(weight_t)
-
-        self._weight_fp4 = weight_fp4
-        self._weight_descale = weight_descale
-        self._fp4_prepared = True
-
-    @torch.no_grad()
-    def _prepare_fp8_sm90_weights(self):
-        """Prepare weights for FP8_SM90 backend (fp8_blockscale_gemm_sm90).
-
-        Uses 128x128 block-scale quantization for optimal SM90 performance.
-        """
-        if self._fp8_sm90_prepared:
-            return
-
-        weight = self.weight.data  # (N, K)
-        n, k = weight.shape
-
-        # Pad to multiples of 128 for block quantization
-        n_pad = ((n + 127) // 128) * 128
-        k_pad = ((k + 127) // 128) * 128
-
-        weight_padded = torch.zeros(
-            n_pad, k_pad, dtype=weight.dtype, device=weight.device
-        )
-        weight_padded[:n, :k] = weight
-
-        # Reshape for 128x128 block quantization
-        # Shape: (N_blocks, 128, K_blocks, 128)
-        weight_view = weight_padded.view(n_pad // 128, 128, k_pad // 128, 128)
-
-        # Compute per-block max for scaling
-        block_amax = (
-            weight_view.abs().float().amax(dim=(1, 3), keepdim=True).clamp(min=1e-4)
-        )
-
-        # FP8 E4M3 max value is ~448
-        fp8_max = 448.0
-        scale = block_amax / fp8_max
-
-        # Quantize to FP8
-        weight_scaled = (weight_view / scale).to(torch.float8_e4m3fn)
-
-        # Store unpadded weight and scales
-        self._weight_fp8_sm90 = weight_scaled.view(n_pad, k_pad)[:n, :k].contiguous()
-        # Scales shape: (N // 128, K // 128) for 128x128 blocks
-        self._weight_scale_sm90 = scale.view(n_pad // 128, k_pad // 128)[
-            : ((n + 127) // 128), : ((k + 127) // 128)
-        ].contiguous()
-        self._fp8_sm90_prepared = True
-
-    @torch.no_grad()
-    def _prepare_bmm_fp8_weights(self):
-        """Prepare weights for BMM_FP8 backend (bmm_fp8 with cublas).
-
-        Uses per-tensor scaling for FP8 quantization.
-        """
-        if self._bmm_fp8_prepared:
-            return
-
-        weight = self.weight.data  # (N, K)
-
-        # Per-tensor FP8 quantization
-        finfo = torch.finfo(torch.float8_e4m3fn)
-        amax = weight.abs().max().clamp(min=1e-12)
-        scale = finfo.max / amax
-
-        # Quantize to FP8
-        weight_scaled = (weight * scale).clamp(finfo.min, finfo.max)
-        weight_fp8 = weight_scaled.to(torch.float8_e4m3fn)
-
-        # bmm_fp8 expects B in (batch, K, N) column-major format
-        # Create as (1, N, K) then transpose to get column-major (1, K, N)
-        # DO NOT call .contiguous() - column-major means non-contiguous transpose view
-        self._weight_fp8_bmm = weight_fp8.unsqueeze(0).transpose(
-            -2, -1
-        )  # (1, K, N) column-major
-        self._weight_scale_bmm = (1.0 / scale).to(torch.float32)
-        self._bmm_fp8_prepared = True
-
-    @torch.no_grad()
-    def _prepare_fp8_nt_weights(self):
-        """Prepare weights for FP8_GROUPWISE / FP8_BLOCKSCALED / BATCH_DEEPGEMM_FP8.
-
-        Quantizes weight (N, K) to FP8 with 128x128 block scales.
-        The NT format keeps weight as (N, K) — no transpose needed.
-        """
-        if self._fp8_nt_prepared:
-            return
-
-        weight = self.weight.data  # (N, K)
-        n, k = weight.shape
-
-        # Block quantization with 128x128 tiles
-        block_size = 128
-        n_pad = ((n + block_size - 1) // block_size) * block_size
-        k_pad = ((k + block_size - 1) // block_size) * block_size
-
-        weight_padded = torch.zeros(
-            n_pad, k_pad, dtype=weight.dtype, device=weight.device
-        )
-        weight_padded[:n, :k] = weight
-
-        weight_view = weight_padded.view(
-            n_pad // block_size, block_size, k_pad // block_size, block_size
-        )
-
-        block_amax = (
-            weight_view.abs().float().amax(dim=(1, 3), keepdim=True).clamp(min=1e-4)
-        )
-
-        fp8_max = 448.0
-        scale = block_amax / fp8_max
-
-        weight_scaled = (weight_view / scale).to(torch.float8_e4m3fn)
-
-        self._weight_fp8_nt = weight_scaled.view(n_pad, k_pad)[:n, :k].contiguous()
-        # scale shape: (n_blocks, k_blocks) stored as (k_blocks, n_blocks) for MN-major
-        raw_scale = scale.squeeze(1).squeeze(-1)  # (n_blocks, k_blocks)
-        n_blocks = (n + block_size - 1) // block_size
-        k_blocks = (k + block_size - 1) // block_size
-        self._weight_scale_nt = (
-            raw_scale[:n_blocks, :k_blocks].t().contiguous().to(torch.float32)
-        )
-        self._fp8_nt_prepared = True
-
-    @torch.no_grad()
-    def _prepare_bmm_bf16_weights(self):
-        """Prepare weights for BMM_BF16 backend (bmm_bf16).
-
-        Pre-transposes weight to column-major layout for bmm_bf16.
-        """
-        # bmm_bf16 expects B in (batch, K, N) column-major format
-        # Our weight is (N, K), create as (1, N, K) then transpose to (1, K, N)
-        # DO NOT call .contiguous() - column-major means non-contiguous transpose view
-        self._weight_t_cache = self.weight.data.unsqueeze(0).transpose(
-            -2, -1
-        )  # (1, K, N) column-major
-
-    @torch.no_grad()
-    def _prepare_mxfp8_weights(self):
-        """Prepare weights for MXFP8 backend (mm_mxfp8).
-
-        Uses FlashInfer's mxfp8_quantize with block size 32.
-        """
-        if self._mxfp8_prepared:
-            return
-
-        weight = self.weight.data  # (N, K)
-        # mm_mxfp8 expects weight in column-major (K, N) format
-        weight_t = weight.t().contiguous()  # (K, N)
-
-        # Quantize weight to MXFP8 using FlashInfer
-        # mxfp8_quantize expects (M, K) layout, so we pass (K, N) as our "M, K"
-        weight_mxfp8, weight_scale = flashinfer.mxfp8_quantize(
-            weight_t, is_sf_swizzled_layout=True
-        )
-
-        self._weight_mxfp8 = weight_mxfp8
-        self._weight_scale_mxfp8 = weight_scale
-        self._mxfp8_prepared = True
-
-    @torch.no_grad()
-    def _prepare_bmm_mxfp8_weights(self):
-        """Prepare weights for BMM_MXFP8 backend (bmm_mxfp8).
-
-        Uses FlashInfer's mxfp8_quantize with block size 32.
-        """
-        if self._bmm_mxfp8_prepared:
-            return
-
-        weight = self.weight.data  # (N, K)
-        # bmm_mxfp8 expects B in (batch, K, N) column-major format
-        weight_t = weight.t().contiguous()  # (K, N)
-
-        # Quantize weight to MXFP8 using FlashInfer
-        weight_mxfp8, weight_scale = flashinfer.mxfp8_quantize(
-            weight_t, is_sf_swizzled_layout=True
-        )
-
-        # Add batch dimension
-        self._weight_mxfp8_bmm = weight_mxfp8.unsqueeze(0)  # (1, K, N)
-        self._weight_scale_mxfp8_bmm = weight_scale.unsqueeze(0)  # (1, ...)
-        self._bmm_mxfp8_prepared = True
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """
-        Forward pass with optimized GEMM.
-
-        Args:
-            x: Input tensor of shape (..., in_features)
-
-        Returns:
-            Output tensor of shape (..., out_features)
-        """
-        # Flatten batch dimensions
-        orig_shape = x.shape
-        x_2d = x.reshape(-1, self.in_features)
-
-        if self._backend == GEMMBackend.TORCH:
-            out = self._forward_torch(x_2d)
-        elif self._backend == GEMMBackend.BF16:
-            out = self._forward_bf16(x_2d)
-        elif self._backend == GEMMBackend.FP8:
-            out = self._forward_fp8(x_2d)
-        elif self._backend == GEMMBackend.FP8_SM90:
-            out = self._forward_fp8_sm90(x_2d)
-        elif self._backend == GEMMBackend.BMM_FP8:
-            out = self._forward_bmm_fp8(x_2d)
-        elif self._backend == GEMMBackend.FP8_GROUPWISE:
-            out = self._forward_fp8_groupwise(x_2d)
-        elif self._backend == GEMMBackend.FP8_BLOCKSCALED:
-            out = self._forward_fp8_blockscaled(x_2d)
-        elif self._backend == GEMMBackend.BATCH_DEEPGEMM_FP8:
-            out = self._forward_batch_deepgemm_fp8(x_2d)
-        elif self._backend == GEMMBackend.FP4:
-            out = self._forward_fp4(x_2d)
-        elif self._backend == GEMMBackend.BMM_BF16:
-            out = self._forward_bmm_bf16(x_2d)
-        elif self._backend == GEMMBackend.MXFP8:
-            out = self._forward_mxfp8(x_2d)
-        elif self._backend == GEMMBackend.BMM_MXFP8:
-            out = self._forward_bmm_mxfp8(x_2d)
-        else:
-            out = self._forward_torch(x_2d)
-
-        # Restore batch dimensions
-        return out.reshape(*orig_shape[:-1], self.out_features)
-
-    def _forward_torch(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using standard PyTorch."""
-        return F.linear(x, self.weight, self.bias)
-
-    def _forward_bf16(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer mm_bf16."""
-        # Ensure BF16
-        x_bf16 = x.to(torch.bfloat16) if x.dtype != torch.bfloat16 else x
-
-        # Cache transposed weight
-        if self._weight_t_cache is None:
-            self._prepare_bf16_weights()
-
-        weight_t = self._weight_t_cache.to(torch.bfloat16)
-
-        # mm_bf16: a @ b where a is (m, k), b is (k, n) column-major
-        out = flashinfer.mm_bf16(x_bf16, weight_t)
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _quantize_activation_fp8_per_tensor(
-        self, x: torch.Tensor
-    ) -> Tuple[torch.Tensor, float]:
-        """Quantize activation to FP8 with per-tensor scale.
-
-        When online_act_quant=True, computes scale from the actual tensor.
-        When online_act_quant=False, uses a fixed default scale (1.0).
-        """
-        finfo = torch.finfo(torch.float8_e4m3fn)
-        if self.online_act_quant:
-            x_amax = x.abs().max().clamp(min=1e-12)
-            x_scale = finfo.max / x_amax
-        else:
-            x_scale = torch.tensor(1.0, device=x.device, dtype=torch.float32)
-        x_fp8 = (x * x_scale).clamp(finfo.min, finfo.max).to(torch.float8_e4m3fn)
-        return x_fp8, x_scale
-
-    def _quantize_activation_fp8_blockwise(
-        self, x: torch.Tensor, block_size: int = 128
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Quantize activation to FP8 with block-wise (per-token-group) scales.
-
-        When online_act_quant=True, computes scales from actual data.
-        When online_act_quant=False, uses a fixed default scale.
-
-        Returns (x_fp8, x_scale) where x_scale shape is (m, k // block_size).
-        """
-        m, k = x.shape
-        fp8_max = 448.0
-        k_pad = ((k + block_size - 1) // block_size) * block_size
-
-        if self.online_act_quant:
-            x_padded = torch.zeros(m, k_pad, dtype=x.dtype, device=x.device)
-            x_padded[:, :k] = x
-            x_view = x_padded.view(m, -1, block_size)
-            x_amax = x_view.abs().float().amax(dim=2).clamp(min=1e-4)
-            x_scale = x_amax / fp8_max  # (m, k_blocks)
-            x_scaled = (x_view / x_scale.unsqueeze(2)).to(torch.float8_e4m3fn)
-            x_fp8 = x_scaled.view(m, k_pad)[:, :k].contiguous()
-        else:
-            default_scale_val = 1.0 / fp8_max
-            x_fp8 = (x * fp8_max).clamp(-fp8_max, fp8_max).to(torch.float8_e4m3fn)
-            num_blocks = k_pad // block_size
-            x_scale = torch.full(
-                (m, num_blocks),
-                default_scale_val,
-                dtype=torch.float32,
-                device=x.device,
-            )
-
-        return x_fp8, x_scale
-
-    def _forward_fp8(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer mm_fp8."""
-        if not self._fp8_prepared:
-            self._prepare_fp8_weights()
-
-        x_fp8, x_scale = self._quantize_activation_fp8_per_tensor(x)
-
-        # Compute output scale
-        alpha = (1.0 / x_scale) * self._weight_scale
-
-        # Run FP8 GEMM
-        out = flashinfer.mm_fp8(x_fp8, self._weight_fp8, alpha=alpha)
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        # Cast back to input dtype for consistency with other backends and to
-        # avoid unexpected upcasts in the surrounding model graph.
-        return out.to(x.dtype)
-
-    def _forward_fp4(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer mm_fp4."""
-        if not self._fp4_prepared:
-            self._prepare_fp4_weights()
-
-        # Quantize input to FP4
-        x_fp4, x_descale = flashinfer.nvfp4_quantize(x.contiguous())
-
-        # Run FP4 GEMM
-        out = flashinfer.mm_fp4(
-            x_fp4,
-            self._weight_fp4,
-            x_descale,
-            self._weight_descale,
-            out_dtype=x.dtype,
-        )
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out
-
-    def _forward_fp8_sm90(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer fp8_blockscale_gemm_sm90.
-
-        This uses the DeepGEMM backend optimized for SM90 (Hopper) GPUs.
-        Supports BF16 input with internal quantization or pre-quantized FP8 weights.
-        """
-        if not self._fp8_sm90_prepared:
-            self._prepare_fp8_sm90_weights()
-
-        # Ensure BF16 input (kernel does internal quantization)
-        x_bf16 = x.to(torch.bfloat16) if x.dtype != torch.bfloat16 else x
-
-        # fp8_blockscale_gemm_sm90: input (M, K), weight (N, K) -> output (M, N)
-        # With FP8 weight and block scales
-        out = flashinfer.gemm.fp8_blockscale_gemm_sm90(
-            x_bf16,
-            self._weight_fp8_sm90,
-            input_scale=None,  # BF16 input, no scale needed
-            weight_scale=self._weight_scale_sm90,
-            out_dtype=torch.bfloat16,
-        )
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_bmm_fp8(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer bmm_fp8 with cublas backend.
-
-        This uses per-tensor FP8 quantization with cublas.
-        """
-        if not self._bmm_fp8_prepared:
-            self._prepare_bmm_fp8_weights()
-
-        x_fp8, x_scale = self._quantize_activation_fp8_per_tensor(x)
-
-        # bmm_fp8 expects (batch, M, K) input
-        x_fp8_batch = x_fp8.unsqueeze(0)  # (1, M, K)
-
-        # Compute scales for output
-        a_scale = (1.0 / x_scale).to(torch.float32).unsqueeze(0)  # (1,)
-        b_scale = self._weight_scale_bmm.unsqueeze(0)  # (1,)
-
-        # Run bmm_fp8: (1, M, K) @ (1, K, N) -> (1, M, N)
-        out = flashinfer.gemm.bmm_fp8(
-            x_fp8_batch,
-            self._weight_fp8_bmm,
-            a_scale,
-            b_scale,
-            dtype=torch.bfloat16,
-            backend="cublas",
-        )
-
-        # Remove batch dimension
-        out = out.squeeze(0)  # (M, N)
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_fp8_groupwise(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer gemm_fp8_nt_groupwise.
-
-        NT GEMM: a(m,k) @ b(n,k)^T with groupwise (1, 128, 128) block scales.
-        The cutlass backend requires a non-None ``scale_major_mode``; we use
-        ``"MN"``, which means scales are laid out with k-block as the leading
-        dimension. ``_weight_scale_nt`` is already prepared in MN-major layout
-        (``(k_blocks, n_blocks)``) by ``_prepare_fp8_nt_weights``; we transpose
-        the per-token activation scale to match.
-        """
-        if not self._fp8_nt_prepared:
-            self._prepare_fp8_nt_weights()
-
-        x_fp8, x_scale = self._quantize_activation_fp8_blockwise(x, block_size=128)
-        # x_scale is (m, k_blocks) K-major; convert to (k_blocks, m) MN-major.
-        x_scale_mn = x_scale.t().contiguous()
-
-        out = gemm_fp8_nt_groupwise(
-            x_fp8,
-            self._weight_fp8_nt,
-            x_scale_mn,
-            self._weight_scale_nt,
-            scale_major_mode="MN",
-            scale_granularity_mnk=(1, 128, 128),
-            out_dtype=torch.bfloat16,
-            backend="cutlass",
-        )
-
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_fp8_blockscaled(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer gemm_fp8_nt_blockscaled.
-
-        Same as groupwise but fixed (128, 128, 128) scale granularity.
-        """
-        if not self._fp8_nt_prepared:
-            self._prepare_fp8_nt_weights()
-
-        x_fp8, x_scale = self._quantize_activation_fp8_blockwise(x, block_size=128)
-
-        # gemm_fp8_nt_blockscaled: wrapper over groupwise with (128,128,128) granularity
-        # a_scale for blockscaled: MN-major -> (k_blocks, m_blocks)
-        m = x.shape[0]
-        block_size = 128
-        m_blocks = (m + block_size - 1) // block_size
-        k_blocks = x_scale.shape[1]
-
-        # Reshape per-token scale to per-block scale for 128x128 blocking
-        x_scale_padded = torch.zeros(
-            m_blocks * block_size,
-            k_blocks,
-            dtype=torch.float32,
-            device=x.device,
-        )
-        x_scale_padded[:m] = x_scale
-        x_scale_block = (
-            x_scale_padded.view(m_blocks, block_size, k_blocks).amax(
-                dim=1
-            )  # (m_blocks, k_blocks)
-        )
-        # MN-major: (k_blocks, m_blocks)
-        a_scale_mn = x_scale_block.t().contiguous()
-
-        out = gemm_fp8_nt_blockscaled(
-            x_fp8,
-            self._weight_fp8_nt,
-            a_scale_mn,
-            self._weight_scale_nt,
-            scale_major_mode="MN",
-            out_dtype=torch.bfloat16,
-        )
-
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_batch_deepgemm_fp8(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer batch_deepgemm_fp8_nt_groupwise.
-
-        Treats the linear as a single-group batched GEMM (batch_size=1).
-        """
-        if not self._fp8_nt_prepared:
-            self._prepare_fp8_nt_weights()
-
-        m = x.shape[0]
-        x_fp8, x_scale = self._quantize_activation_fp8_blockwise(x, block_size=128)
-
-        # batch_deepgemm expects (batch, m, k) inputs
-        a = x_fp8.unsqueeze(0)  # (1, m, k)
-        b = self._weight_fp8_nt.unsqueeze(0)  # (1, n, k)
-        a_scale_batch = x_scale.unsqueeze(0)  # (1, m, k_blocks)
-        b_scale_batch = self._weight_scale_nt.unsqueeze(
-            0
-        )  # (1, k_blocks, n_blocks) -> need (1, n_blocks, k_blocks)
-        # batch_deepgemm b_scale expects (batch, n_blocks, k_blocks)
-        b_scale_batch = b_scale_batch.transpose(1, 2).contiguous()
-        masked_m = torch.tensor([m], dtype=torch.int32, device=x.device)
-
-        out = batch_deepgemm_fp8_nt_groupwise(
-            a,
-            b,
-            a_scale_batch,
-            b_scale_batch,
-            masked_m=masked_m,
-            expected_m=m,
-            scale_granularity_mnk=(1, 128, 128),
-            out_dtype=torch.bfloat16,
-        )
-
-        out = out.squeeze(0)  # (m, n)
-
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_bmm_bf16(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer bmm_bf16.
-
-        Uses batched matrix multiply with BF16 precision.
-        """
-        # Prepare weights if not already done
-        if self._weight_t_cache is None:
-            self._prepare_bmm_bf16_weights()
-
-        # Ensure BF16
-        x_bf16 = x.to(torch.bfloat16) if x.dtype != torch.bfloat16 else x
-
-        # bmm_bf16 expects (batch, M, K) input
-        x_batch = x_bf16.unsqueeze(0)  # (1, M, K)
-        weight_bf16 = self._weight_t_cache.to(torch.bfloat16)
-
-        # Run bmm_bf16: (1, M, K) @ (1, K, N) -> (1, M, N)
-        out = flashinfer.bmm_bf16(x_batch, weight_bf16)
-
-        # Remove batch dimension
-        out = out.squeeze(0)  # (M, N)
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_mxfp8(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer mm_mxfp8.
-
-        Uses MXFP8 quantization with block size 32.
-        """
-        if not self._mxfp8_prepared:
-            self._prepare_mxfp8_weights()
-
-        # Quantize input to MXFP8
-        x_mxfp8, x_scale = flashinfer.mxfp8_quantize(
-            x.contiguous(), is_sf_swizzled_layout=True
-        )
-
-        # Run mm_mxfp8: (M, K) @ (K, N) -> (M, N)
-        out = flashinfer.mm_mxfp8(
-            x_mxfp8,
-            self._weight_mxfp8,
-            x_scale,
-            self._weight_scale_mxfp8,
-            out_dtype=torch.bfloat16,
-        )
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    def _forward_bmm_mxfp8(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward using FlashInfer bmm_mxfp8.
-
-        Uses batched MXFP8 quantization with block size 32.
-        """
-        if not self._bmm_mxfp8_prepared:
-            self._prepare_bmm_mxfp8_weights()
-
-        # Quantize input to MXFP8
-        x_mxfp8, x_scale = flashinfer.mxfp8_quantize(
-            x.contiguous(), is_sf_swizzled_layout=True
-        )
-
-        # bmm_mxfp8 expects (batch, M, K) input
-        x_batch = x_mxfp8.unsqueeze(0)  # (1, M, K)
-        x_scale_batch = x_scale.unsqueeze(0)  # (1, ...)
-
-        # Run bmm_mxfp8: (1, M, K) @ (1, K, N) -> (1, M, N)
-        out = flashinfer.bmm_mxfp8(
-            x_batch,
-            self._weight_mxfp8_bmm,
-            x_scale_batch,
-            self._weight_scale_mxfp8_bmm,
-            dtype=torch.bfloat16,
-        )
-
-        # Remove batch dimension
-        out = out.squeeze(0)  # (M, N)
-
-        # Add bias if present
-        if self.bias is not None:
-            out = out + self.bias.to(out.dtype)
-
-        return out.to(x.dtype)
-
-    @property
-    def backend(self) -> str:
-        """Return current backend name."""
-        return self._backend.value
-
-    def extra_repr(self) -> str:
-        return (
-            f"in_features={self.in_features}, out_features={self.out_features}, "
-            f"bias={self.has_bias}, backend={self._backend.value}"
-        )
+_EXAMPLES_PYTORCH_DIR = Path(__file__).resolve().parents[1]
+if str(_EXAMPLES_PYTORCH_DIR) not in sys.path:
+    sys.path.insert(0, str(_EXAMPLES_PYTORCH_DIR))
+
+from flashinfer_modules import (
+    GEMMBackend,
+    FlashInferAttentionDispatcher,
+    FlashInferFP32LayerNorm,
+    FlashInferFeedForward,
+    FlashInferLinear,
+    FlashInferRMSNorm,
+    apply_rotary_emb,
+    create_linear_layer,
+    get_1d_rotary_pos_embed,
+)
 
 
 @dataclass
@@ -1191,8 +66,7 @@ class WanTransformer3DConfig:
     Added fields for FlashInfer optimizations:
     - gemm_backend: GEMM backend for linear layers (see GEMMBackend enum, default "torch")
     - online_act_quant: True for online activation scale computation, False for fixed default scale
-    - single_attention_backend: "single", "cudnn", or "trtllm" — used when batch_size == 1
-    - batch_attention_backend: "cudnn" or "trtllm" — used when batch_size > 1
+    - attention_backend: "auto", "single", "cudnn", or "trtllm"
     - use_skip_softmax_sparse: Whether to use skip-softmax sparse attention (trtllm backend)
     - skip_softmax_threshold_scale_factor: Threshold scale factor for skip-softmax sparsity
     """
@@ -1232,12 +106,7 @@ class WanTransformer3DConfig:
     online_act_quant: bool = (
         True  # True: compute activation scale from data; False: use default scale
     )
-    single_attention_backend: Literal["single", "cudnn", "trtllm"] = (
-        "single"  # Attention backend for batch_size == 1
-    )
-    batch_attention_backend: Literal["cudnn", "trtllm"] = (
-        "cudnn"  # Attention backend for batch_size > 1
-    )
+    attention_backend: Literal["auto", "single", "cudnn", "trtllm"] = "auto"
     use_skip_softmax_sparse: bool = (
         False  # Enable skip-softmax sparse attention (trtllm only)
     )
@@ -1252,231 +121,61 @@ class WanTransformer3DConfig:
         return self.num_attention_heads * self.attention_head_dim
 
 
-def create_linear_layer(
-    in_features: int,
-    out_features: int,
-    bias: bool = True,
-    gemm_backend: str = "torch",
-    device: Optional[torch.device] = None,
-    dtype: Optional[torch.dtype] = None,
-    online_act_quant: bool = True,
-) -> nn.Module:
-    """
-    Factory function to create a linear layer with the specified GEMM backend.
-
-    If gemm_backend is "torch", returns standard nn.Linear.
-    Otherwise, returns FlashInferLinear with the specified backend.
-    """
-    if gemm_backend == "torch":
-        return nn.Linear(
-            in_features, out_features, bias=bias, device=device, dtype=dtype
-        )
-    else:
-        return FlashInferLinear(
-            in_features,
-            out_features,
-            bias=bias,
-            backend=gemm_backend,
-            device=device,
-            dtype=dtype,
-            online_act_quant=online_act_quant,
-        )
+_FLASHINFER_ENV_OVERRIDES = {
+    "gemm_backend": "FLASHINFER_GEMM_BACKEND",
+    "online_act_quant": "FLASHINFER_ONLINE_ACT_QUANT",
+    "use_skip_softmax_sparse": "FLASHINFER_USE_SKIP_SOFTMAX_SPARSE",
+    "skip_softmax_threshold_scale_factor": "FLASHINFER_SKIP_SOFTMAX_THRESHOLD",
+}
 
 
-def get_1d_rotary_pos_embed(
-    dim: int,
-    pos: int,
-    theta: float = 10000.0,
-    use_real: bool = True,
-    repeat_interleave_real: bool = True,
-    freqs_dtype: torch.dtype = torch.float64,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """Generate 1D rotary position embeddings."""
-    freqs = 1.0 / (theta ** (torch.arange(0, dim, 2, dtype=freqs_dtype) / dim))
-    t = torch.arange(pos, dtype=freqs_dtype)
-    freqs = torch.outer(t, freqs)
-
-    if use_real and repeat_interleave_real:
-        freqs_cos = freqs.cos().repeat_interleave(2, dim=-1)
-        freqs_sin = freqs.sin().repeat_interleave(2, dim=-1)
-    else:
-        freqs_cos = freqs.cos()
-        freqs_sin = freqs.sin()
-
-    return freqs_cos.float(), freqs_sin.float()
+def _env_bool(name: str) -> Optional[bool]:
+    value = os.getenv(name)
+    if value is None:
+        return None
+    value = value.strip().lower()
+    if value in {"1", "true", "yes", "on"}:
+        return True
+    if value in {"0", "false", "no", "off"}:
+        return False
+    raise ValueError(
+        f"{name} must be a boolean value (1/0, true/false, yes/no, on/off), got {value!r}."
+    )
 
 
-class FlashInferRMSNorm(nn.Module):
-    """RMSNorm using FlashInfer kernel."""
+def _config_value(config: Any, name: str, default: Any) -> Any:
+    if isinstance(config, dict):
+        return config.get(name, default)
+    return getattr(config, name, default)
 
-    def __init__(
-        self, hidden_size: int, eps: float = 1e-6, elementwise_affine: bool = True
-    ):
-        super().__init__()
-        self.eps = eps
-        self.elementwise_affine = elementwise_affine
-        if elementwise_affine:
-            self.weight = nn.Parameter(torch.ones(hidden_size))
+
+def _config_to_flashinfer_config(
+    config: Any, overrides: Optional[dict] = None
+) -> WanTransformer3DConfig:
+    overrides = dict(overrides or {})
+    values = {
+        field: _config_value(config, field, getattr(WanTransformer3DConfig, field))
+        for field in WanTransformer3DConfig.__dataclass_fields__
+        if field != "inner_dim"
+    }
+
+    attention_backend = os.getenv("FLASHINFER_ATTENTION_BACKEND")
+    if attention_backend:
+        values["attention_backend"] = attention_backend
+
+    for field, env_name in _FLASHINFER_ENV_OVERRIDES.items():
+        raw = os.getenv(env_name)
+        if raw is None:
+            continue
+        if field in {"online_act_quant", "use_skip_softmax_sparse"}:
+            values[field] = _env_bool(env_name)
+        elif field == "skip_softmax_threshold_scale_factor":
+            values[field] = float(raw)
         else:
-            self.register_parameter("weight", None)
+            values[field] = raw
 
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        if self.weight is not None:
-            # FlashInfer rmsnorm only supports FP16/BF16, cast if needed
-            if x.dtype == torch.float32:
-                x_bf16 = x.to(torch.bfloat16)
-                weight_bf16 = self.weight.to(torch.bfloat16)
-                out = flashinfer.rmsnorm(x_bf16.contiguous(), weight_bf16, self.eps)
-                return out.to(torch.float32)
-            return flashinfer.rmsnorm(x.contiguous(), self.weight, self.eps)
-        else:
-            # Fallback for non-affine case
-            variance = x.pow(2).mean(-1, keepdim=True)
-            return x * torch.rsqrt(variance + self.eps)
-
-
-class FlashInferFP32LayerNorm(nn.Module):
-    """LayerNorm computed in FP32 using FlashInfer kernel when applicable."""
-
-    def __init__(
-        self,
-        normalized_shape: int,
-        eps: float = 1e-5,
-        elementwise_affine: bool = True,
-    ):
-        super().__init__()
-        self.normalized_shape = normalized_shape
-        self.eps = eps
-        self.elementwise_affine = elementwise_affine
-        if elementwise_affine:
-            self.weight = nn.Parameter(torch.ones(normalized_shape))
-            self.bias = nn.Parameter(torch.zeros(normalized_shape))
-        else:
-            self.register_parameter("weight", None)
-            self.register_parameter("bias", None)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        orig_dtype = x.dtype
-        x_float = x.float()
-
-        if self.elementwise_affine and x.dim() == 2:
-            # Use FlashInfer layernorm for 2D tensors with affine params
-            # FlashInfer layernorm expects weight/bias in float32
-            weight = (
-                self.weight.float()
-                if self.weight is not None
-                else torch.ones(
-                    self.normalized_shape, device=x.device, dtype=torch.float32
-                )
-            )
-            bias = (
-                self.bias.float()
-                if self.bias is not None
-                else torch.zeros(
-                    self.normalized_shape, device=x.device, dtype=torch.float32
-                )
-            )
-            out = flashinfer.layernorm(
-                x_float.to(torch.bfloat16), weight, bias, self.eps
-            )
-            return out.to(orig_dtype)
-        else:
-            # Fallback for 3D+ tensors or non-affine
-            out = F.layer_norm(
-                x_float,
-                (self.normalized_shape,),
-                self.weight.float() if self.weight is not None else None,
-                self.bias.float() if self.bias is not None else None,
-                self.eps,
-            )
-            return out.to(orig_dtype)
-
-
-class FlashInferFeedForward(nn.Module):
-    """Feed-forward network with GELU activation and FlashInfer GEMM support.
-
-    Note: The original Wan model uses simple GELU without gating (up->gelu->down).
-    FlashInfer's gelu_and_mul is for gated FFN (gelu(gate) * up), which is different.
-    We use PyTorch's F.gelu here since FlashInfer doesn't have a standalone GELU kernel.
-
-    Args:
-        dim: Input/output dimension
-        inner_dim: Hidden dimension
-        activation_fn: Activation function ("gelu-approximate", "gelu")
-        bias: Whether to use bias in linear layers
-        use_gating: Whether to use gated FFN (gelu(x) * gate)
-        gemm_backend: GEMM backend for linear layers (see GEMMBackend enum)
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        inner_dim: int,
-        activation_fn: str = "gelu-approximate",
-        bias: bool = True,
-        use_gating: bool = False,  # Whether to use gated FFN (gelu(x) * gate)
-        gemm_backend: str = "torch",
-        online_act_quant: bool = True,
-    ):
-        super().__init__()
-        self.activation_fn = activation_fn
-        self.use_gating = use_gating
-        self.gemm_backend = gemm_backend
-
-        if use_gating:
-            self.proj_up = create_linear_layer(
-                dim,
-                2 * inner_dim,
-                bias=bias,
-                gemm_backend=gemm_backend,
-                online_act_quant=online_act_quant,
-            )
-        else:
-            self.proj_up = create_linear_layer(
-                dim,
-                inner_dim,
-                bias=bias,
-                gemm_backend=gemm_backend,
-                online_act_quant=online_act_quant,
-            )
-
-        self.proj_down = create_linear_layer(
-            inner_dim,
-            dim,
-            bias=bias,
-            gemm_backend=gemm_backend,
-            online_act_quant=online_act_quant,
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        if self.use_gating:
-            # Gated FFN using FlashInfer fused gelu_and_mul
-            hidden = self.proj_up(x)
-
-            # Check alignment requirement (16-byte aligned)
-            if hidden.shape[-1] * hidden.dtype.itemsize % 16 == 0:
-                # FlashInfer activation only supports FP16/BF16
-                orig_dtype = hidden.dtype
-                if orig_dtype == torch.float32:
-                    hidden = hidden.to(torch.bfloat16)
-
-                if self.activation_fn == "gelu-approximate":
-                    hidden = flashinfer.gelu_tanh_and_mul(hidden.contiguous())
-                else:
-                    hidden = flashinfer.gelu_and_mul(hidden.contiguous())
-
-                if orig_dtype == torch.float32:
-                    hidden = hidden.to(orig_dtype)
-            else:
-                # Fallback for non-aligned
-                gate, up = hidden.chunk(2, dim=-1)
-                hidden = F.gelu(gate, approximate="tanh") * up
-        else:
-            # Simple FFN: up -> activation -> down
-            hidden = self.proj_up(x)
-            hidden = F.gelu(hidden, approximate="tanh")
-
-        return self.proj_down(hidden)
+    values.update({k: v for k, v in overrides.items() if v is not None})
+    return WanTransformer3DConfig(**values)
 
 
 class WanRotaryPosEmbed(nn.Module):
@@ -1557,39 +256,8 @@ class WanRotaryPosEmbed(nn.Module):
         return freqs_cos, freqs_sin
 
 
-def apply_rotary_emb(
-    hidden_states: torch.Tensor,
-    freqs_cos: torch.Tensor,
-    freqs_sin: torch.Tensor,
-) -> torch.Tensor:
-    """Apply rotary position embedding to hidden states."""
-    x1, x2 = hidden_states.unflatten(-1, (-1, 2)).unbind(-1)
-    cos = freqs_cos[..., 0::2]
-    sin = freqs_sin[..., 1::2]
-    out = torch.empty_like(hidden_states)
-    out[..., 0::2] = x1 * cos - x2 * sin
-    out[..., 1::2] = x1 * sin + x2 * cos
-    return out.type_as(hidden_states)
-
-
 class FlashInferWanAttention(nn.Module):
-    """Attention module using FlashInfer kernels with optimized GEMM backends.
-
-    Args:
-        dim: Input dimension
-        heads: Number of attention heads
-        dim_head: Dimension per head
-        eps: Epsilon for normalization
-        dropout: Dropout rate
-        added_kv_proj_dim: Dimension for additional KV projection (I2V)
-        cross_attention_dim_head: Head dimension for cross attention
-        is_cross_attention: Whether this is cross attention
-        gemm_backend: GEMM backend for linear layers (see GEMMBackend enum)
-        use_skip_softmax_sparse: Whether to use skip-softmax sparse attention
-        skip_softmax_threshold_scale_factor: Threshold scale factor for skip-softmax sparsity
-        text_encoder_context_length: Context length of the text encoder; used to
-            split image vs text tokens in I2V cross-attention.
-    """
+    """Wan attention module that delegates FlashInfer kernel dispatch."""
 
     def __init__(
         self,
@@ -1603,43 +271,27 @@ class FlashInferWanAttention(nn.Module):
         is_cross_attention: Optional[bool] = None,
         gemm_backend: str = "torch",
         online_act_quant: bool = True,
-        single_attention_backend: str = "single",
-        batch_attention_backend: str = "cudnn",
+        attention_backend: str = "auto",
         use_skip_softmax_sparse: bool = False,
         skip_softmax_threshold_scale_factor: float = 1.0,
         text_encoder_context_length: int = 512,
     ):
         super().__init__()
-
         self.inner_dim = dim_head * heads
         self.heads = heads
         self.dim_head = dim_head
-        self.added_kv_proj_dim = added_kv_proj_dim
-        self.cross_attention_dim_head = cross_attention_dim_head
         self.kv_inner_dim = (
             self.inner_dim
             if cross_attention_dim_head is None
             else cross_attention_dim_head * heads
         )
-        self.gemm_backend = gemm_backend
-        self.single_attention_backend = single_attention_backend
-        self.batch_attention_backend = batch_attention_backend
-        self.use_skip_softmax_sparse = use_skip_softmax_sparse
-        self.skip_softmax_threshold_scale_factor = skip_softmax_threshold_scale_factor
         self.text_encoder_context_length = text_encoder_context_length
-
-        # Lazily-allocated scratch buffers reused across forward passes. These
-        # are deliberately NOT registered as torch buffers — they are not part
-        # of the model state, just per-device workspaces for FlashInfer kernels.
-        # Pre-allocating once avoids the ~tens-of-ms overhead of repeatedly
-        # zeroing a 128 MB tensor inside ``forward`` (which dominates latency
-        # for sparse attention on every layer of the transformer).
-        self._workspace_buffer: Optional[torch.Tensor] = None
-        self._ragged_prefill_wrapper: Optional[BatchPrefillWithRaggedKVCacheWrapper] = (
-            None
+        self.is_cross_attention = (
+            is_cross_attention
+            if is_cross_attention is not None
+            else cross_attention_dim_head is not None
         )
 
-        # Use FlashInfer-optimized linear layers
         self.to_q = create_linear_layer(
             dim,
             self.inner_dim,
@@ -1674,7 +326,6 @@ class FlashInferWanAttention(nn.Module):
             ]
         )
 
-        # Use FlashInfer RMSNorm for Q/K normalization
         self.norm_q = FlashInferRMSNorm(
             dim_head * heads, eps=eps, elementwise_affine=True
         )
@@ -1702,313 +353,13 @@ class FlashInferWanAttention(nn.Module):
                 dim_head * heads, eps=eps, elementwise_affine=False
             )
 
-        if is_cross_attention is not None:
-            self.is_cross_attention = is_cross_attention
-        else:
-            self.is_cross_attention = cross_attention_dim_head is not None
-
-    def _get_workspace_buffer(self, device: torch.device) -> torch.Tensor:
-        """Return a per-device 128 MB workspace buffer, allocated once.
-
-        The buffer is reused by ``cudnn_batch_prefill``, ``trtllm_batch_context``
-        and the ragged-prefill wrapper. ``trtllm_batch_context_with_kv_cache``
-        only requires the workspace to be zeroed on first use, so we allocate
-        with ``zeros`` and never reset it afterwards.
-        """
-        if self._workspace_buffer is None or self._workspace_buffer.device != device:
-            self._workspace_buffer = torch.zeros(
-                _DEFAULT_WORKSPACE_SIZE, dtype=torch.uint8, device=device
-            )
-            # Recreating the workspace invalidates any pre-bound wrapper that
-            # captured the old buffer pointer.
-            self._ragged_prefill_wrapper = None
-        return self._workspace_buffer
-
-    def _get_ragged_prefill_wrapper(
-        self, device: torch.device
-    ) -> BatchPrefillWithRaggedKVCacheWrapper:
-        """Return a cached ``BatchPrefillWithRaggedKVCacheWrapper``.
-
-        The wrapper caches its own internal data structures across ``plan``
-        calls; ``plan`` is still cheap to call per-forward (the heavy work is
-        in ``run``).
-        """
-        workspace = self._get_workspace_buffer(device)
-        if self._ragged_prefill_wrapper is None:
-            self._ragged_prefill_wrapper = BatchPrefillWithRaggedKVCacheWrapper(
-                workspace, kv_layout="NHD"
-            )
-        return self._ragged_prefill_wrapper
-
-    def _resolve_attention_backend(
-        self, batch_size: int, device: torch.device
-    ) -> Tuple[str, bool]:
-        """Resolve the attention backend and whether to use sparse attention.
-
-        Returns
-        -------
-        backend : str
-            One of ``"single"``, ``"cudnn"``, ``"trtllm"``.
-        use_sparse : bool
-            ``True`` only when sparse skip-softmax was requested *and* the
-            current GPU supports it. When ``False`` the caller must NOT pass
-            ``skip_softmax_threshold_scale_factor`` to the kernel — otherwise
-            a kernel that does not support sparse attention (e.g. TRT-LLM on
-            non-Blackwell) would silently produce wrong results or crash.
-        """
-        # Default: the configured backend for the given batch size, no sparse.
-        backend = (
-            self.single_attention_backend
-            if batch_size == 1
-            else self.batch_attention_backend
+        self.attention_dispatcher = FlashInferAttentionDispatcher(
+            heads=heads,
+            dim_head=dim_head,
+            attention_backend=attention_backend,
+            use_skip_softmax_sparse=use_skip_softmax_sparse,
+            skip_softmax_threshold_scale_factor=skip_softmax_threshold_scale_factor,
         )
-        use_sparse = False
-
-        if self.use_skip_softmax_sparse:
-            major, minor = get_compute_capability(device)
-            sm_version = major * 10 + minor
-            if sm_version in (100, 103):
-                # Skip-softmax sparse attention is only implemented in the
-                # TRT-LLM kernel on SM100/103.
-                return "trtllm", True
-            warnings.warn(
-                f"Skip-softmax sparse attention requires SM100 or SM103 "
-                f"(Blackwell B100/B200), but current GPU is SM{sm_version}. "
-                f"Falling back to {backend!r} with standard (non-sparse) attention.",
-                RuntimeWarning,
-                stacklevel=2,
-            )
-
-        return backend, use_sparse
-
-    def _attention_single(
-        self,
-        query: torch.Tensor,
-        key: torch.Tensor,
-        value: torch.Tensor,
-        batch_size: int,
-        seq_len_q: int,
-        seq_len_kv: int,
-    ) -> torch.Tensor:
-        """Batched ragged prefill via ``BatchPrefillWithRaggedKVCacheWrapper``.
-
-        Replaces the original per-batch ``single_prefill_with_kv_cache`` loop:
-        FlashInfer's ragged-KV wrapper handles the entire batch in a single
-        kernel launch, eliminating per-sequence Python and launch overhead.
-        For ``batch_size == 1`` performance is comparable to ``single_prefill``;
-        for larger batches the speedup grows with the number of sequences.
-        """
-        device = query.device
-        wrapper = self._get_ragged_prefill_wrapper(device)
-
-        qo_indptr = torch.arange(
-            0,
-            (batch_size + 1) * seq_len_q,
-            seq_len_q,
-            dtype=torch.int32,
-            device=device,
-        )
-        kv_indptr = torch.arange(
-            0,
-            (batch_size + 1) * seq_len_kv,
-            seq_len_kv,
-            dtype=torch.int32,
-            device=device,
-        )
-
-        wrapper.plan(
-            qo_indptr,
-            kv_indptr,
-            num_qo_heads=self.heads,
-            num_kv_heads=self.heads,
-            head_dim_qk=self.dim_head,
-            causal=False,
-            q_data_type=query.dtype,
-        )
-
-        # Wrapper expects (total_tokens, heads, head_dim) ragged layout.
-        query_flat = query.reshape(
-            batch_size * seq_len_q, self.heads, self.dim_head
-        ).contiguous()
-        key_flat = key.reshape(
-            batch_size * seq_len_kv, self.heads, self.dim_head
-        ).contiguous()
-        value_flat = value.reshape(
-            batch_size * seq_len_kv, self.heads, self.dim_head
-        ).contiguous()
-
-        out = wrapper.run(query_flat, key_flat, value_flat)
-        return out.view(batch_size, seq_len_q, -1)
-
-    def _attention_cudnn_batch(
-        self,
-        query: torch.Tensor,
-        key: torch.Tensor,
-        value: torch.Tensor,
-        batch_size: int,
-        seq_len_q: int,
-        seq_len_kv: int,
-    ) -> torch.Tensor:
-        """Batch attention using cudnn_batch_prefill_with_kv_cache."""
-        device = query.device
-
-        # Flatten query: (batch, seq_q, heads, head_dim) -> (total_tokens, heads, head_dim)
-        query_flat = query.reshape(
-            batch_size * seq_len_q, self.heads, self.dim_head
-        ).contiguous()
-        # Non-paged KV: (total_kv_tokens, heads_kv, head_dim)
-        k_flat = key.reshape(
-            batch_size * seq_len_kv, self.heads, self.dim_head
-        ).contiguous()
-        v_flat = value.reshape(
-            batch_size * seq_len_kv, self.heads, self.dim_head
-        ).contiguous()
-
-        actual_seq_lens_q = torch.full(
-            (batch_size,), seq_len_q, dtype=torch.int32, device=device
-        )
-        actual_seq_lens_kv = torch.full(
-            (batch_size,), seq_len_kv, dtype=torch.int32, device=device
-        )
-
-        workspace = self._get_workspace_buffer(device)
-
-        sm_scale = 1.0 / math.sqrt(self.dim_head)
-
-        # Compute batch offsets for ragged layout
-        batch_offsets_q = torch.arange(
-            0, batch_size * seq_len_q, seq_len_q, dtype=torch.int32, device=device
-        )
-        batch_offsets_kv = torch.arange(
-            0, batch_size * seq_len_kv, seq_len_kv, dtype=torch.int32, device=device
-        )
-
-        out, _ = cudnn_batch_prefill_with_kv_cache(
-            q=query_flat,
-            k_cache=k_flat,
-            v_cache=v_flat,
-            scale=sm_scale,
-            workspace_buffer=workspace,
-            max_token_per_sequence=seq_len_q,
-            max_sequence_kv=seq_len_kv,
-            actual_seq_lens_q=actual_seq_lens_q,
-            actual_seq_lens_kv=actual_seq_lens_kv,
-            causal=False,
-            return_lse=False,
-            batch_offsets_q=batch_offsets_q,
-            batch_offsets_o=batch_offsets_q,
-            batch_offsets_k=batch_offsets_kv,
-            batch_offsets_v=batch_offsets_kv,
-        )
-
-        return out.view(batch_size, seq_len_q, -1)
-
-    def _attention_trtllm_batch(
-        self,
-        query: torch.Tensor,
-        key: torch.Tensor,
-        value: torch.Tensor,
-        batch_size: int,
-        seq_len_q: int,
-        seq_len_kv: int,
-        use_sparse: bool = False,
-    ) -> torch.Tensor:
-        """Batch attention using trtllm_batch_context_with_kv_cache.
-
-        ``use_sparse=True`` enables skip-softmax sparse attention, which
-        currently only works on SM100/SM103. The caller is responsible for
-        gating this via :meth:`_resolve_attention_backend`.
-        """
-        device = query.device
-
-        # Set up paged KV cache: [num_pages, 2, num_heads, page_size, head_dim] in HND layout
-        key_paged = key.permute(0, 2, 1, 3).contiguous()
-        value_paged = value.permute(0, 2, 1, 3).contiguous()
-        kv_cache = torch.stack([key_paged, value_paged], dim=1)
-
-        block_tables = torch.arange(
-            batch_size, dtype=torch.int32, device=device
-        ).unsqueeze(1)
-
-        seq_lens = torch.full(
-            (batch_size,), seq_len_kv, dtype=torch.int32, device=device
-        )
-
-        cum_seq_lens_q = torch.arange(
-            0,
-            (batch_size + 1) * seq_len_q,
-            seq_len_q,
-            dtype=torch.int32,
-            device=device,
-        )
-        cum_seq_lens_kv = torch.arange(
-            0,
-            (batch_size + 1) * seq_len_kv,
-            seq_len_kv,
-            dtype=torch.int32,
-            device=device,
-        )
-
-        workspace = self._get_workspace_buffer(device)
-
-        query_flat = query.reshape(
-            batch_size * seq_len_q, self.heads, self.dim_head
-        ).contiguous()
-
-        sm_scale = 1.0 / math.sqrt(self.dim_head)
-
-        skip_threshold = (
-            self.skip_softmax_threshold_scale_factor if use_sparse else None
-        )
-
-        hidden_states = trtllm_batch_context_with_kv_cache(
-            query=query_flat,
-            kv_cache=kv_cache,
-            workspace_buffer=workspace,
-            block_tables=block_tables,
-            seq_lens=seq_lens,
-            max_q_len=seq_len_q,
-            max_kv_len=seq_len_kv,
-            bmm1_scale=sm_scale,
-            bmm2_scale=1.0,
-            batch_size=batch_size,
-            cum_seq_lens_q=cum_seq_lens_q,
-            cum_seq_lens_kv=cum_seq_lens_kv,
-            kv_layout="HND",
-            skip_softmax_threshold_scale_factor=skip_threshold,
-        )
-        return hidden_states.view(batch_size, seq_len_q, -1)
-
-    def _dispatch_attention(
-        self,
-        backend: str,
-        query: torch.Tensor,
-        key: torch.Tensor,
-        value: torch.Tensor,
-        batch_size: int,
-        seq_len_q: int,
-        seq_len_kv: int,
-        use_sparse: bool = False,
-    ) -> torch.Tensor:
-        """Dispatch to the appropriate attention implementation.
-
-        ``use_sparse`` is only honored by the trtllm path; for other backends
-        we silently ignore it (callers should not request sparse on backends
-        that don't support it — :meth:`_resolve_attention_backend` enforces
-        this).
-        """
-        if backend == "single":
-            return self._attention_single(
-                query, key, value, batch_size, seq_len_q, seq_len_kv
-            )
-        elif backend == "trtllm":
-            return self._attention_trtllm_batch(
-                query, key, value, batch_size, seq_len_q, seq_len_kv, use_sparse
-            )
-        else:  # cudnn
-            return self._attention_cudnn_batch(
-                query, key, value, batch_size, seq_len_q, seq_len_kv
-            )
 
     def forward(
         self,
@@ -2019,8 +370,6 @@ class FlashInferWanAttention(nn.Module):
     ) -> torch.Tensor:
         encoder_hidden_states_img = None
         if self.add_k_proj is not None and encoder_hidden_states is not None:
-            # I2V models prepend image tokens to the text encoder output. Split
-            # them apart using the configured text-encoder context length.
             image_context_length = (
                 encoder_hidden_states.shape[1] - self.text_encoder_context_length
             )
@@ -2030,15 +379,10 @@ class FlashInferWanAttention(nn.Module):
         if encoder_hidden_states is None:
             encoder_hidden_states = hidden_states
 
-        query = self.to_q(hidden_states)
-        key = self.to_k(encoder_hidden_states)
+        query = self.norm_q(self.to_q(hidden_states))
+        key = self.norm_k(self.to_k(encoder_hidden_states))
         value = self.to_v(encoder_hidden_states)
 
-        # Apply RMSNorm to Q and K using FlashInfer
-        query = self.norm_q(query)
-        key = self.norm_k(key)
-
-        # Reshape to (batch, seq, heads, head_dim)
         batch_size = hidden_states.shape[0]
         seq_len_q = hidden_states.shape[1]
         seq_len_kv = encoder_hidden_states.shape[1]
@@ -2047,12 +391,10 @@ class FlashInferWanAttention(nn.Module):
         key = key.view(batch_size, seq_len_kv, self.heads, self.dim_head)
         value = value.view(batch_size, seq_len_kv, self.heads, self.dim_head)
 
-        # Apply rotary embedding
         if rotary_emb is not None:
             query = apply_rotary_emb(query, *rotary_emb)
             key = apply_rotary_emb(key, *rotary_emb)
 
-        # FlashInfer attention only supports FP16/BF16, cast if needed
         orig_dtype = query.dtype
         needs_cast = orig_dtype == torch.float32
         if needs_cast:
@@ -2060,7 +402,6 @@ class FlashInferWanAttention(nn.Module):
             key = key.to(torch.bfloat16)
             value = value.to(torch.bfloat16)
 
-        # I2V task: handle image cross-attention
         hidden_states_img = None
         if encoder_hidden_states_img is not None:
             key_img = self.add_k_proj(encoder_hidden_states_img)
@@ -2076,10 +417,10 @@ class FlashInferWanAttention(nn.Module):
                 key_img = key_img.to(torch.bfloat16)
                 value_img = value_img.to(torch.bfloat16)
 
-            # Image cross-attention: never sparse — it operates on a separate
-            # set of (small) image tokens and must always run dense.
-            img_backend, _ = self._resolve_attention_backend(batch_size, query.device)
-            hidden_states_img = self._dispatch_attention(
+            img_backend, _ = self.attention_dispatcher._resolve_attention_backend(
+                batch_size, query.device
+            )
+            hidden_states_img = self.attention_dispatcher._dispatch_attention(
                 img_backend,
                 query,
                 key_img,
@@ -2090,11 +431,10 @@ class FlashInferWanAttention(nn.Module):
                 use_sparse=False,
             )
 
-        # Main attention
-        attn_backend, use_sparse = self._resolve_attention_backend(
+        attn_backend, use_sparse = self.attention_dispatcher._resolve_attention_backend(
             batch_size, query.device
         )
-        hidden_states = self._dispatch_attention(
+        hidden_states = self.attention_dispatcher._dispatch_attention(
             attn_backend,
             query,
             key,
@@ -2105,7 +445,6 @@ class FlashInferWanAttention(nn.Module):
             use_sparse=use_sparse,
         )
 
-        # Cast back to original dtype if needed
         if needs_cast:
             hidden_states = hidden_states.to(orig_dtype)
             if hidden_states_img is not None:
@@ -2146,8 +485,7 @@ class FlashInferWanTransformerBlock(nn.Module):
         added_kv_proj_dim: Optional[int] = None,
         gemm_backend: str = "torch",
         online_act_quant: bool = True,
-        single_attention_backend: str = "single",
-        batch_attention_backend: str = "cudnn",
+        attention_backend: str = "auto",
         use_skip_softmax_sparse: bool = False,
         skip_softmax_threshold_scale_factor: float = 1.0,
         text_encoder_context_length: int = 512,
@@ -2164,8 +502,7 @@ class FlashInferWanTransformerBlock(nn.Module):
             cross_attention_dim_head=None,
             gemm_backend=gemm_backend,
             online_act_quant=online_act_quant,
-            single_attention_backend=single_attention_backend,
-            batch_attention_backend=batch_attention_backend,
+            attention_backend=attention_backend,
             use_skip_softmax_sparse=use_skip_softmax_sparse,
             skip_softmax_threshold_scale_factor=skip_softmax_threshold_scale_factor,
             text_encoder_context_length=text_encoder_context_length,
@@ -2181,8 +518,7 @@ class FlashInferWanTransformerBlock(nn.Module):
             cross_attention_dim_head=dim // num_heads,
             gemm_backend=gemm_backend,
             online_act_quant=online_act_quant,
-            single_attention_backend=single_attention_backend,
-            batch_attention_backend=batch_attention_backend,
+            attention_backend=attention_backend,
             use_skip_softmax_sparse=False,  # No sparse attention for cross-attention
             text_encoder_context_length=text_encoder_context_length,
         )
@@ -2498,11 +834,10 @@ class FlashInferWanTransformer3DModel(nn.Module):
     def __init__(self, config: Union[WanTransformer3DConfig, Any]):
         super().__init__()
 
-        # Handle both dataclass config and HuggingFace config
-        if hasattr(config, "patch_size"):
-            self.config = config
-        else:
-            self.config = WanTransformer3DConfig()
+        # Normalize HuggingFace/diffusers configs to the local dataclass and
+        # apply environment overrides once at construction time.
+        config = _config_to_flashinfer_config(config)
+        self.config = config
 
         patch_size = getattr(config, "patch_size", (1, 2, 2))
         num_attention_heads = getattr(config, "num_attention_heads", 40)
@@ -2524,8 +859,7 @@ class FlashInferWanTransformer3DModel(nn.Module):
         # FlashInfer optimization options
         gemm_backend = getattr(config, "gemm_backend", "torch")
         online_act_quant = getattr(config, "online_act_quant", True)
-        single_attention_backend = getattr(config, "single_attention_backend", "single")
-        batch_attention_backend = getattr(config, "batch_attention_backend", "cudnn")
+        attention_backend = getattr(config, "attention_backend", "auto")
         use_skip_softmax_sparse = getattr(config, "use_skip_softmax_sparse", False)
         skip_softmax_threshold_scale_factor = getattr(
             config, "skip_softmax_threshold_scale_factor", 1.0
@@ -2540,8 +874,7 @@ class FlashInferWanTransformer3DModel(nn.Module):
         self.patch_size = patch_size
         self.gemm_backend = gemm_backend
         self.online_act_quant = online_act_quant
-        self.single_attention_backend = single_attention_backend
-        self.batch_attention_backend = batch_attention_backend
+        self.attention_backend = attention_backend
         self.use_skip_softmax_sparse = use_skip_softmax_sparse
         self.skip_softmax_threshold_scale_factor = skip_softmax_threshold_scale_factor
 
@@ -2576,8 +909,7 @@ class FlashInferWanTransformer3DModel(nn.Module):
                     added_kv_proj_dim,
                     gemm_backend=gemm_backend,
                     online_act_quant=online_act_quant,
-                    single_attention_backend=single_attention_backend,
-                    batch_attention_backend=batch_attention_backend,
+                    attention_backend=attention_backend,
                     use_skip_softmax_sparse=use_skip_softmax_sparse,
                     skip_softmax_threshold_scale_factor=skip_softmax_threshold_scale_factor,
                     text_encoder_context_length=text_encoder_context_length,
@@ -2612,6 +944,17 @@ class FlashInferWanTransformer3DModel(nn.Module):
             if isinstance(module, FlashInferLinear):
                 module.prepare_weights()
 
+    @property
+    def dtype(self) -> torch.dtype:
+        try:
+            return next(self.parameters()).dtype
+        except StopIteration:
+            return torch.float32
+
+    def cache_context(self, _name: str):
+        """Compatibility shim for diffusers pipelines that cache cond/uncond calls."""
+        return nullcontext()
+
     def forward(
         self,
         hidden_states: torch.Tensor,
@@ -2619,6 +962,7 @@ class FlashInferWanTransformer3DModel(nn.Module):
         encoder_hidden_states: torch.Tensor,
         encoder_hidden_states_image: Optional[torch.Tensor] = None,
         return_dict: bool = True,
+        attention_kwargs: Optional[dict] = None,
     ) -> Union[torch.Tensor, dict]:
         """
         Forward pass of the FlashInfer-optimized Wan Transformer.
@@ -2732,8 +1076,15 @@ class FlashInferWanTransformer3DModel(nn.Module):
         Note: This requires the original diffusers model to be installed
         for config loading.
         """
+        flashinfer_kwargs = {}
+        for name in WanTransformer3DConfig.__dataclass_fields__:
+            if name in kwargs:
+                flashinfer_kwargs[name] = kwargs.pop(name)
+
         try:
-            from diffusers import WanTransformer3DModel as OriginalModel
+            from diffusers.models.transformers.transformer_wan import (
+                WanTransformer3DModel as OriginalModel,
+            )
         except ImportError as e:
             raise ImportError(
                 "Please install diffusers to load from pretrained: "
@@ -2744,7 +1095,8 @@ class FlashInferWanTransformer3DModel(nn.Module):
         original_model = OriginalModel.from_pretrained(model_path, **kwargs)
 
         # Create FlashInfer model with same config
-        model = cls(original_model.config)
+        config = _config_to_flashinfer_config(original_model.config, flashinfer_kwargs)
+        model = cls(config)
 
         # Copy weights. ``strict=False`` is required because some FlashInfer
         # GEMM backends (FP8, FP4, MXFP8, ...) register additional buffers
@@ -2771,12 +1123,36 @@ class FlashInferWanTransformer3DModel(nn.Module):
 
 if __name__ == "__main__":
     import argparse
+    import time
 
-    parser = argparse.ArgumentParser(description="Test FlashInfer WAN Transformer")
+    parser = argparse.ArgumentParser(
+        description="Load a Hugging Face Wan transformer checkpoint and run FlashInfer forward."
+    )
+    parser.add_argument(
+        "--model-id",
+        type=str,
+        default="Wan-AI/Wan2.1-T2V-1.3B-Diffusers",
+        help="Hugging Face repo id or local checkpoint path.",
+    )
+    parser.add_argument(
+        "--subfolder",
+        type=str,
+        default="transformer",
+        help="Subfolder containing the WanTransformer3DModel checkpoint.",
+    )
+    parser.add_argument("--revision", type=str, default=None)
+    parser.add_argument("--variant", type=str, default=None)
+    parser.add_argument(
+        "--dtype",
+        type=str,
+        default="bfloat16",
+        choices=["float16", "bfloat16", "float32"],
+        help="Model and input dtype.",
+    )
     parser.add_argument(
         "--gemm-backend",
         type=str,
-        default="torch",
+        default=os.getenv("FLASHINFER_GEMM_BACKEND", "torch"),
         choices=[b.value for b in GEMMBackend],
         help="GEMM backend for linear layers",
     )
@@ -2786,18 +1162,11 @@ if __name__ == "__main__":
         help="Use offline (fixed default) activation quantization instead of online",
     )
     parser.add_argument(
-        "--single-attention-backend",
+        "--attention-backend",
         type=str,
-        default="single",
-        choices=["single", "cudnn", "trtllm"],
-        help="Attention backend for batch_size == 1",
-    )
-    parser.add_argument(
-        "--batch-attention-backend",
-        type=str,
-        default="cudnn",
-        choices=["cudnn", "trtllm"],
-        help="Attention backend for batch_size > 1",
+        default=os.getenv("FLASHINFER_ATTENTION_BACKEND", "auto"),
+        choices=["auto", "single", "cudnn", "trtllm"],
+        help="Attention backend: auto uses single for bs=1 and cudnn for bs>1",
     )
     parser.add_argument(
         "--skip-softmax-sparse",
@@ -2810,40 +1179,65 @@ if __name__ == "__main__":
         default=1.0,
         help="Threshold scale factor for skip-softmax (higher = more sparse)",
     )
+    parser.add_argument("--batch-size", type=int, default=1)
+    parser.add_argument("--num-frames", type=int, default=4)
+    parser.add_argument("--height", type=int, default=32)
+    parser.add_argument("--width", type=int, default=32)
+    parser.add_argument(
+        "--text-seq-len",
+        type=int,
+        default=512,
+        help="Synthetic encoder_hidden_states sequence length.",
+    )
+    parser.add_argument("--warmup-iters", type=int, default=1)
+    parser.add_argument("--benchmark-iters", type=int, default=1)
+    parser.add_argument(
+        "--prepare-weights",
+        action="store_true",
+        help="Pre-convert FlashInferLinear weights for the selected GEMM backend.",
+    )
     args = parser.parse_args()
 
-    print("Testing FlashInferWanTransformer3DModel...")
-    print(f"GEMM Backend: {args.gemm_backend}")
-    print(f"Online Act Quant: {not args.offline_act_quant}")
-    print(f"Single Attention Backend: {args.single_attention_backend}")
-    print(f"Batch Attention Backend: {args.batch_attention_backend}")
-    print(f"Skip-Softmax Sparse: {args.skip_softmax_sparse}")
-    if args.skip_softmax_sparse:
-        print(f"Threshold Scale Factor: {args.skip_softmax_threshold}")
+    if args.benchmark_iters < 1:
+        raise ValueError("--benchmark-iters must be >= 1.")
+    if not torch.cuda.is_available():
+        raise RuntimeError("CUDA is required to run this FlashInfer example.")
 
-    config = WanTransformer3DConfig(
-        patch_size=(1, 2, 2),
-        num_attention_heads=8,
-        attention_head_dim=64,
-        in_channels=4,
-        out_channels=4,
-        text_dim=512,
-        freq_dim=256,
-        ffn_dim=1024,
-        num_layers=2,
-        cross_attn_norm=True,
-        eps=1e-6,
-        rope_max_seq_len=256,
-        # FlashInfer optimization options
-        gemm_backend=args.gemm_backend,
-        online_act_quant=not args.offline_act_quant,
-        single_attention_backend=args.single_attention_backend,
-        batch_attention_backend=args.batch_attention_backend,
-        use_skip_softmax_sparse=args.skip_softmax_sparse,
-        skip_softmax_threshold_scale_factor=args.skip_softmax_threshold,
+    dtype = {
+        "float16": torch.float16,
+        "bfloat16": torch.bfloat16,
+        "float32": torch.float32,
+    }[args.dtype]
+
+    print(f"Loading checkpoint: {args.model_id} subfolder={args.subfolder!r}")
+    print(
+        "FlashInfer config: "
+        f"gemm_backend={args.gemm_backend}, "
+        f"online_act_quant={not args.offline_act_quant}, "
+        f"attention_backend={args.attention_backend}, "
+        f"use_skip_softmax_sparse={args.skip_softmax_sparse}"
     )
 
-    model = FlashInferWanTransformer3DModel(config).cuda().half()
+    load_kwargs = {
+        "subfolder": args.subfolder,
+        "torch_dtype": dtype,
+        "gemm_backend": args.gemm_backend,
+        "online_act_quant": not args.offline_act_quant,
+        "attention_backend": args.attention_backend,
+        "use_skip_softmax_sparse": args.skip_softmax_sparse,
+        "skip_softmax_threshold_scale_factor": args.skip_softmax_threshold,
+    }
+    if args.revision is not None:
+        load_kwargs["revision"] = args.revision
+    if args.variant is not None:
+        load_kwargs["variant"] = args.variant
+
+    model = FlashInferWanTransformer3DModel.from_pretrained(
+        args.model_id, **load_kwargs
+    )
+    model = model.to(device="cuda", dtype=dtype).eval()
+    if args.prepare_weights:
+        model.prepare_weights()
 
     # Print model info
     linear_count = sum(1 for m in model.modules() if isinstance(m, FlashInferLinear))
@@ -2851,48 +1245,40 @@ if __name__ == "__main__":
     print(f"FlashInferLinear layers: {linear_count}")
     print(f"Standard nn.Linear layers: {torch_linear_count}")
 
-    # Test inputs
-    batch_size = 1
-    num_frames = 4
-    height = 32
-    width = 32
-    in_channels = 4
-    text_seq_len = 64
+    in_channels = model.config.in_channels
+    text_dim = model.config.text_dim
 
     hidden_states = torch.randn(
-        batch_size,
+        args.batch_size,
         in_channels,
-        num_frames,
-        height,
-        width,
+        args.num_frames,
+        args.height,
+        args.width,
         device="cuda",
-        dtype=torch.float16,
+        dtype=dtype,
     )
-    timestep = torch.randint(0, 1000, (batch_size,), device="cuda")
+    timestep = torch.randint(0, 1000, (args.batch_size,), device="cuda")
     encoder_hidden_states = torch.randn(
-        batch_size, text_seq_len, 512, device="cuda", dtype=torch.float16
+        args.batch_size, args.text_seq_len, text_dim, device="cuda", dtype=dtype
     )
 
-    # Warmup
     with torch.no_grad():
-        for _ in range(3):
+        for _ in range(args.warmup_iters):
             _ = model(hidden_states, timestep, encoder_hidden_states, return_dict=False)
-
-    # Benchmark
-    import time
 
     torch.cuda.synchronize()
     start = time.time()
-    num_iters = 10
     with torch.no_grad():
-        for _ in range(num_iters):
+        for _ in range(args.benchmark_iters):
             output = model(
                 hidden_states, timestep, encoder_hidden_states, return_dict=False
             )
-            torch.cuda.synchronize()
+    torch.cuda.synchronize()
     elapsed = time.time() - start
 
     print(f"Input shape: {hidden_states.shape}")
     print(f"Output shape: {output[0].shape}")
-    print(f"Average time per forward pass: {elapsed / num_iters * 1000:.2f} ms")
-    print("Test passed!")
+    print(
+        f"Average time per forward pass: "
+        f"{elapsed / args.benchmark_iters * 1000:.2f} ms"
+    )
